chore: simplify decimal input test, fix linux locale handling
